webapp/requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
python-dotenv
django-quill-editor==0.1.*
